[Remove] メタ情報系のヘルパーを別モジュールに分離
